+++ src/game/tests.rs
@@ -771,6 +771,7 @@ mod urza_lands;
 mod venarian_gold;
 
 mod vintage_cube_abhorrent_oculus;
+mod vintage_cube_adeline;
 mod vintage_cube_adventure;
 mod vintage_cube_agathas_soul_cauldron;
 mod vintage_cube_ajani;